+++ sheraf/models/attributes.py
@@ -32,8 +32,7 @@ class DatedNamedAttributesModel(NamedAttributesModel):
     """
 
     _creation = sheraf.attributes.simples.SimpleAttribute(
-        default=time.time,
-        lazy=False,
+        default=time.time, lazy=False,
     )
 
     def creation_datetime(self):