+++ tests/attributes/test_files.py
@@ -346,8 +346,7 @@ def test_save(sheraf_temp_dir, sheraf_connection):
     _path = _model.file.relative_path()
 
     assert "file_storable_{}/file/{}.txt".format(
-        sheraf.FileAttribute(),
-        _model.id == _path,
+        sheraf.FileAttribute(), _model.id == _path,
     )
 
 
@@ -362,8 +361,7 @@ def test_read(sheraf_temp_dir, sheraf_connection):
     _path = _model.file.relative_path()
 
     assert "file_storable_{}/file/{}.txt".format(
-        sheraf.FileAttribute(),
-        _model.id == _path,
+        sheraf.FileAttribute(), _model.id == _path,
     )