+++ tests/models/test_indexation.py
@@ -57,12 +57,7 @@ class UniqueModelC(tests.IntAutoModel):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        UniqueModelA,
-        UniqueModelB,
-        UniqueModelC,
-    ],
+    "Model", [UniqueModelA, UniqueModelB, UniqueModelC,],
 )
 def test_unique_index_creation(sheraf_database, Model):
     assert (
@@ -94,12 +89,7 @@ def test_unique_index_creation(sheraf_database, Model):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        UniqueModelA,
-        UniqueModelB,
-        UniqueModelC,
-    ],
+    "Model", [UniqueModelA, UniqueModelB, UniqueModelC,],
 )
 def test_unique_index_creation_and_edition(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -130,12 +120,7 @@ def test_unique_index_creation_and_edition(sheraf_database, Model):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        UniqueModelA,
-        UniqueModelB,
-        UniqueModelC,
-    ],
+    "Model", [UniqueModelA, UniqueModelB, UniqueModelC,],
 )
 def test_unique_index_creation_and_deletion(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -153,12 +138,7 @@ def test_unique_index_creation_and_deletion(sheraf_database, Model):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        UniqueModelA,
-        UniqueModelB,
-        UniqueModelC,
-    ],
+    "Model", [UniqueModelA, UniqueModelB, UniqueModelC,],
 )
 def test_unique_index_creation_and_attribute_deletion(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -182,12 +162,7 @@ def test_unique_index_creation_and_attribute_deletion(sheraf_database, Model):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        UniqueModelA,
-        UniqueModelB,
-        UniqueModelC,
-    ],
+    "Model", [UniqueModelA, UniqueModelB, UniqueModelC,],
 )
 def test_unique_index_double_value(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -250,12 +225,7 @@ class MultipleModelC(tests.IntAutoModel):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        MultipleModelA,
-        MultipleModelB,
-        MultipleModelC,
-    ],
+    "Model", [MultipleModelA, MultipleModelB, MultipleModelC,],
 )
 def test_multiple_index_creation(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -283,12 +253,7 @@ def test_multiple_index_creation(sheraf_database, Model):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        MultipleModelA,
-        MultipleModelB,
-        MultipleModelC,
-    ],
+    "Model", [MultipleModelA, MultipleModelB, MultipleModelC,],
 )
 def test_multiple_index_creation_and_deletion(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -381,11 +346,7 @@ class MultipleKeysIndexModelC(tests.IntAutoModel):
 
 @pytest.mark.parametrize(
     "Model",
-    [
-        MultipleKeysIndexModelA,
-        MultipleKeysIndexModelB,
-        MultipleKeysIndexModelC,
-    ],
+    [MultipleKeysIndexModelA, MultipleKeysIndexModelB, MultipleKeysIndexModelC,],
 )
 def test_multiple_keys_index_create(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -422,11 +383,7 @@ def test_multiple_keys_index_create(sheraf_database, Model):
 
 
 @pytest.mark.parametrize(
-    "Model",
-    [
-        MultipleKeysIndexModelA,
-        MultipleKeysIndexModelB,
-    ],
+    "Model", [MultipleKeysIndexModelA, MultipleKeysIndexModelB,],
 )
 def test_multiple_keys_index_update(sheraf_database, Model):
     with sheraf.connection(commit=True):
@@ -610,10 +567,7 @@ class CustomSearchModelC(tests.IntAutoModel):
     bar = sheraf.SimpleAttribute()
 
     fooindex = sheraf.Index(
-        "foo",
-        key="foo",
-        unique=True,
-        values=lambda string: {string.lower()},
+        "foo", key="foo", unique=True, values=lambda string: {string.lower()},
     )
 
     barindex = sheraf.Index("bar", key="bar")
@@ -628,10 +582,7 @@ class CustomSearchModelD(tests.IntAutoModel):
     bar = sheraf.SimpleAttribute()
 
     fooindex = sheraf.Index(
-        "foo",
-        key="foo",
-        unique=True,
-        values=lambda string: {string.lower()},
+        "foo", key="foo", unique=True, values=lambda string: {string.lower()},
     )
 
     barindex = sheraf.Index("bar", key="bar")
@@ -646,10 +597,7 @@ class CustomSearchModelE(tests.IntAutoModel):
     bar = sheraf.SimpleAttribute()
 
     fooindex = sheraf.Index(
-        foo,
-        key="foo",
-        unique=True,
-        values=lambda string: {string.lower()},
+        foo, key="foo", unique=True, values=lambda string: {string.lower()},
     )
 
     barindex = sheraf.Index(bar, key="bar")