+++ tests/tools/test_persistent_reference_dict_merge.py
@@ -6,38 +6,22 @@ from ZODB.ConflictResolution import PersistentReference as PR
 
 def test_edited_once():
     assert {"ref": PR("bar")} == merge(
-        {"ref": PR("foo")},
-        {"ref": PR("foo")},
-        {"ref": PR("bar")},
+        {"ref": PR("foo")}, {"ref": PR("foo")}, {"ref": PR("bar")},
     )
 
     assert {"ref": PR("bar")} == merge(
-        {"ref": PR("foo")},
-        {"ref": PR("bar")},
-        {"ref": PR("foo")},
+        {"ref": PR("foo")}, {"ref": PR("bar")}, {"ref": PR("foo")},
     )
 
 
 def test_added_once():
-    assert {"ref": PR("foo")} == merge(
-        {},
-        {"ref": PR("foo")},
-        {},
-    )
+    assert {"ref": PR("foo")} == merge({}, {"ref": PR("foo")}, {},)
 
-    assert {"ref": PR("foo")} == merge(
-        {},
-        {},
-        {"ref": PR("foo")},
-    )
+    assert {"ref": PR("foo")} == merge({}, {}, {"ref": PR("foo")},)
 
 
 def test_deleted_twice():
-    assert {} == merge(
-        {"ref": PR("foo")},
-        {},
-        {},
-    )
+    assert {} == merge({"ref": PR("foo")}, {}, {},)
 
 
 def test_added_twice():