+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Fixed
 - :class:`~sheraf.attributes.index.Index` inheritance is fixed.
 - :class:`~sheraf.attributes.index.Index` update :class:`~sheraf.attributes.base.BaseAttribute`
   on deletion.
+- :meth:`~sheraf.databases.connection` can take a ``reuse`` parameter.
+
 
 [0.4.0] - 2021-02-17
 ====================